deprecate old pkd tests with old processing class
