+++ tests/parkfield/test_process_parkfield_run.py
@@ -1,93 +1,87 @@
-from deprecated import deprecated
 from pathlib import Path
-import pickle
 
-from aurora.config.processing_config import RunConfig
+from aurora.config import BANDS_DEFAULT_FILE
+from aurora.config.config_creator import ConfigCreator
+from aurora.config.metadata import Processing
 
-from aurora.pipelines.process_mth5 import process_mth5_run
+from aurora.pipelines.process_mth5_new import process_mth5
 from aurora.test_utils.parkfield.path_helpers import AURORA_RESULTS_PATH
+from aurora.test_utils.parkfield.path_helpers import CONFIG_PATH
+from aurora.test_utils.parkfield.path_helpers import DATA_PATH
 from aurora.test_utils.parkfield.path_helpers import EMTF_RESULTS_PATH
+from aurora.tf_kernel.dataset import DatasetDefinition
+from aurora.tf_kernel.helpers import extract_run_summaries_from_mth5s
 from aurora.transfer_function.plot.comparison_plots import compare_two_z_files
 
-
 from make_parkfield_mth5 import test_make_parkfield_mth5
-from make_processing_configs import create_run_test_config
 
-@deprecated(version="0.0.3", reason="new mt_metadata based config")
-def test_process_to_tf_collection(z_file_path=None):
-    processing_run_cfg = create_run_test_config()
-    print(f"CONFIG {processing_run_cfg.name}")
-
-    config = RunConfig()
-    config.from_json(processing_run_cfg)
-    mth5_path = Path(config.mth5_path)
-
-    # Ensure there is an mth5 to process
-    if not mth5_path.exists():
-        test_make_parkfield_mth5()
-
-    run_id = "001"
-    show_plot = False
-    tf_collection = process_mth5_run(
-        processing_run_cfg,
-        run_id,
-        mth5_path=mth5_path,
-        units="MT",
-        show_plot=show_plot,
-        z_file_path=z_file_path,
-    )
-    return tf_collection
-
-
-@deprecated(version="0.0.3", reason="new mt_metadata based config")
-def test_processing(z_file_path=None):
+DEBUG_ISSUE_172 = False
+def test_processing(return_collection=False, z_file_path=None):
     """
-
     Parameters
     ----------
-    z_file_path
+    return_collection: bool
+        Controls dtype of returned object
+    z_file_path: str or Path or None
+        Where to store zfile
 
     Returns
     -------
-    tf_cls: mt_metadata.transfer_functions.core.TF
+    tf_cls: TF object,
+        if  return_collection is True:
+        aurora.transfer_function.transfer_function_collection.TransferFunctionCollection
+        if  return_collection is False:
+        mt_metadata.transfer_functions.core.TF
     """
-    processing_run_cfg = create_run_test_config()
-    print(f"CONFIG {processing_run_cfg.name}")
 
-    config = RunConfig()
-    config.from_json(processing_run_cfg)
-    mth5_path = Path(config.mth5_path)
+    mth5_path = DATA_PATH.joinpath("pkd_test_00.h5")
 
     # Ensure there is an mth5 to process
     if not mth5_path.exists():
         test_make_parkfield_mth5()
 
-    run_id = "001"
+    run_summary = extract_run_summaries_from_mth5s([mth5_path,])
+    run_summary["remote"] = False
+    cc = ConfigCreator(config_path=CONFIG_PATH)
+    p = cc.create_run_processing_object(emtf_band_file=BANDS_DEFAULT_FILE,
+                                        sample_rate=40.0
+                                        )
+    p.stations.from_dataset_dataframe(run_summary)
+    for decimation in p.decimations:
+        decimation.estimator.engine = "RME"
+
+    if DEBUG_ISSUE_172:
+        config = Processing()
+        config.from_json(processing_run_cfg)
+    else:
+        config = p
+
+    dataset_definition = DatasetDefinition()
+    dataset_definition.df = run_summary
     show_plot = False
+    tf_cls = process_mth5(config,
+                          dataset_definition,
+                         units="MT",
+                         show_plot=show_plot,
+                         z_file_path=z_file_path,
+                         return_collection=return_collection
+                         )
 
-    tf_cls = process_mth5_run(
-        processing_run_cfg,
-        run_id,
-        mth5_path=mth5_path,
-        units="MT",
-        show_plot=show_plot,
-        z_file_path=z_file_path,
-        return_collection=False,
-    )
-
-    # with open(z_file_path.parent.joinpath("tf_cls.pkl"), "wb") as fid:
-    #     pickle.dump(tf_cls2, fid)
-    #     print(f"Pickled tf_cls2 to {z_file_path.parent.joinpath('tf_cls.pkl')}")
-        
-    tf_cls.write_tf_file(fn="emtfxml_test.xml", file_type="emtfxml")
+    if return_collection:
+        tf_collection = tf_cls
+        return tf_collection
+    else:
+        tf_cls.write_tf_file(fn="emtfxml_test.xml", file_type="emtfxml")
     return tf_cls
 
 
-@deprecated(version="0.0.3", reason="new mt_metadata based config")
+
+
+
 def main():
     z_file_path = AURORA_RESULTS_PATH.joinpath("pkd.zss")
-    test_process_to_tf_collection(z_file_path)
-    test_processing(z_file_path)
+    test_processing(return_collection=True, z_file_path=z_file_path)
+    test_processing(return_collection=False, z_file_path=z_file_path)
 
     # COMPARE WITH ARCHIVED Z-FILE
     auxilliary_z_file = EMTF_RESULTS_PATH.joinpath("PKD_272_00.zrr")
